requirements.txt
telethon==1.32.1
requests==2.31.0
python-dotenv==1.0.1
pysocks==1.7.1 
python-socks==2.7.1